Refactored blocks overview component and its import
