+++ src/modules/block/components/blocks/blocks.js
@@ -8,8 +8,8 @@ import FilterBar from 'src/modules/common/components/filterBar';
 import StickyHeader from 'src/theme/table/stickyHeader';
 import Table from 'src/theme/table';
 import { transformStringDateToUnixTimestamp } from 'src/utils/dateTime';
-import BlocksOverview from '../../manager/blocksOverviewManager';
 import { useBlocks } from '../../hooks/queries/useBlocks';
+import BlocksOverview from '../blocksOverview';
 import BlockFilterDropdown from './blockFilterDropdown';
 import BlockRow from './blockRow';
 import header from './tableHeader';
